+++ minishell.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   minishell.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: yeoshin <yeoshin@student.42.fr>            +#+  +:+       +#+        */
+/*   By: soljeong <soljeong@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/29 19:05:52 by soljeong          #+#    #+#             */
-/*   Updated: 2024/04/17 15:02:54 by yeoshin          ###   ########.fr       */
+/*   Updated: 2024/04/19 11:03:36 by soljeong         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,6 +21,7 @@
 # include "parse/parse_list.h"
 # include "parse/parse.h"
 # include "parse/parse_test.h"
+# include "parse/wildcard.h"
 # include "pipe/pipex.h"
 # include "execute_builtin/execute_builtin.h"
 # include "get_next_line/get_next_line_bonus.h"
