+++ parse/tree_utils.c
@@ -6,7 +6,7 @@
 /*   By: soljeong <soljeong@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/02 11:12:37 by soljeong          #+#    #+#             */
-/*   Updated: 2024/04/17 20:21:14 by soljeong         ###   ########.fr       */
+/*   Updated: 2024/04/19 11:10:40 by soljeong         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,12 +22,6 @@ void	clear_tree(t_tree *tree)
 		clear_tree(tree->right);
 	if (tree->token)
 		ft_del_token_node(tree->token);
-	// if (tree->redirect)
-	// {
-	// 	if (tree->redirect->filename)
-	// 		free(tree->redirect->filename);
-	// 	free(tree->redirect);
-	// }
 	if (tree)
 		free(tree);
 }
